ansible/roles/search_index/files/requirements.txt
requests
pika
pyyaml
elasticsearch
